API query to Etherscan.io now working.

+++ src/main/kotlin/com/mpcrypto/ethereumgaschecker/gasscanner/GasSnapshop.kt
@@ -4,6 +4,6 @@ package com.mpcrypto.ethereumgaschecker.gasscanner
  * Data class to simply interactions with ValuesList of GasHistory.
  */
 data class GasSnapshot(
-    val gasValue : Double,
+    val gasValue : Int,
     val timeStamp : Long
 )
